+++ programs/benches/fib_1k.c
@@ -4,7 +4,7 @@
 
 typedef struct fib_return_values
 {
-    uint64_t remaining_gas;
+    __uint128_t remaining_gas;
     struct {
         uint8_t discriminant;
         union {
